requirements.txt
py-cord
python-dotenv
bs4
requests
lxml